projects/fhir-medications-processor/requirements.txt
annotated-types==0.7.0
boto3==1.37.12
botocore==1.37.12
fhir-core==1.0.0
fhir.resources==8.0.0
jmespath==1.0.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
